TIP-1185: Move specs from PQB to Abstract
